refactored buttons
