+++ src/components/Buttons/ClickToOpen.js
@@ -2,7 +2,7 @@
 import React, {useState} from "react"
 import { Button} from "react-bootstrap";
 import styled from "styled-components";
-import MyModal from "./MyModal";
+import MyModal from "../MyModal";
 import "bootstrap/dist/css/bootstrap.min.css";
 
 export default function ClickToOpen(props) {
@@ -10,8 +10,6 @@ export default function ClickToOpen(props) {
     const [show, setShow] = useState(false);
 
     const handleShow = () => setShow(true);
-
-    const handleClose = () => setShow(false);
 
     const MyButton = styled(Button)`
         margin: auto;
@@ -21,7 +19,7 @@ export default function ClickToOpen(props) {
     return (
         <div>
             <MyButton className={props.size} variant={props.variant} onClick={handleShow}>{props.text}</MyButton>
-            <MyModal show={show} setShow={setShow}></MyModal>
+            <MyModal show={show} setShow={setShow} title={props.title} apiUrlFunction={props.apiUrlFunction} apiKey={props.apiKey}></MyModal>
         </div>
     )
 }
